move dimens file to styles folder
