+++ app/(dashboard)/[workspaceId]/[projectId]/page.tsx
@@ -18,7 +18,7 @@ import { useWorkspacesManager } from '@globals/WorkspacesContext';
 // schemas
 import { Project } from '@schemas';
 // styles
-import { project_details_bar_height } from '@defines/dimens';
+import { project_details_bar_height } from '@styles/dimens';
 import styles from './ProjectPage.module.css';
 
 const fallbackDesc = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.';
